Feature/numeric in schemas

+++ benchmarks/tpch/schemas/psql_native_partitioned_date_week/schema.sql
@@ -19,7 +19,7 @@ CREATE TABLE part (
     p_type character varying(25) NOT NULL,
     p_size int NOT NULL,
     p_container character varying(10) NOT NULL,
-    p_retailprice double precision NOT NULL,
+    p_retailprice numeric(13,2) NOT NULL,
     p_comment character varying(23) NOT NULL
 );
 
@@ -29,7 +29,7 @@ CREATE TABLE supplier (
     s_address character varying(40) NOT NULL,
     s_nationkey int NOT NULL,
     s_phone character varying(15) NOT NULL,
-    s_acctbal double precision NOT NULL,
+    s_acctbal numeric(13,2) NOT NULL,
     s_comment character varying(101) NOT NULL
 );
 
@@ -37,7 +37,7 @@ CREATE TABLE partsupp (
     ps_partkey int NOT NULL,
     ps_suppkey int NOT NULL,
     ps_availqty int NOT NULL,
-    ps_supplycost double precision NOT NULL,
+    ps_supplycost numeric(13,2) NOT NULL,
     ps_comment character varying(199) NOT NULL
 );
 
@@ -47,7 +47,7 @@ CREATE TABLE customer (
     c_address character varying(40) NOT NULL,
     c_nationkey int NOT NULL,
     c_phone character varying(15) NOT NULL,
-    c_acctbal double precision NOT NULL,
+    c_acctbal numeric(13,2) NOT NULL,
     c_mktsegment character varying(10) NOT NULL,
     c_comment character varying(117) NOT NULL
 );
@@ -56,7 +56,7 @@ CREATE TABLE orders (
     o_orderkey bigint NOT NULL,
     o_custkey int NOT NULL,
     o_orderstatus "char" NOT NULL,
-    o_totalprice double precision NOT NULL,
+    o_totalprice numeric(13,2) NOT NULL,
     o_orderdate date NOT NULL,
     o_orderpriority character varying(15) NOT NULL,
     o_clerk character varying(15) NOT NULL,
@@ -503,10 +503,10 @@ CREATE TABLE lineitem (
     l_partkey int NOT NULL,
     l_suppkey int NOT NULL,
     l_linenumber int NOT NULL,
-    l_quantity double precision NOT NULL,
-    l_extendedprice double precision NOT NULL,
-    l_discount double precision NOT NULL,
-    l_tax double precision NOT NULL,
+    l_quantity numeric(13,2) NOT NULL,
+    l_extendedprice numeric(13,2) NOT NULL,
+    l_discount numeric(13,2) NOT NULL,
+    l_tax numeric(13,2) NOT NULL,
     l_returnflag "char" NOT NULL,
     l_linestatus "char" NOT NULL,
     l_shipdate date NOT NULL,
